+++ eslint.config.js
@@ -34,5 +34,5 @@ export default tseslint.config(
     },
   },
   eslintPluginPrettier,
-  eslintConfigPrettier
+  eslintConfigPrettier,
 );
